+++ src/interactions/commands/music/play.js
@@ -1,40 +1,42 @@
 /* eslint-disable no-case-declarations */
+const { SlashCommand } = require("@src/structures");
+const { CommandInteraction, MessageEmbed } = require("discord.js");
 const { EMBED_COLORS } = require("@root/config");
-const { Command } = require("@src/structures");
-const { Message, MessageEmbed } = require("discord.js");
 const prettyMs = require("pretty-ms");
 
-module.exports = class Play extends Command {
+module.exports = class Play extends SlashCommand {
   constructor(client) {
     super(client, {
       name: "play",
       description: "play a song from youtube",
-      command: {
-        enabled: true,
-        usage: "<song-name>",
-        minArgsCount: 1,
-        category: "MUSIC",
-        botPermissions: ["EMBED_LINKS"],
-      },
+      enabled: true,
+      options: [
+        {
+          name: "query",
+          description: "song name or url",
+          type: "STRING",
+          required: true,
+        },
+      ],
     });
   }
 
   /**
-   * @param {Message} message
-   * @param {string[]} args
+   * @param {CommandInteraction} interaction
    */
-  async messageRun(message, args) {
-    const { guild, channel, member } = message;
-    const query = args.join(" ");
+  async run(interaction) {
+    const { guild, channel, user } = interaction;
+    const member = await guild.members.fetch(user.id);
+    const query = interaction.options.getString("query");
 
-    if (!member.voice.channel) return channel.send("You need to join a voice channel first");
-    let player = message.client.musicManager.get(guild.id);
+    if (!member.voice.channel) return interaction.followUp("You need to join a voice channel first");
+    let player = guild.client.musicManager.get(guild.id);
 
     if (player && member.voice.channel !== guild.me.voice.channel) {
-      return channel.send("You must be in the same voice channel as mine");
+      return interaction.followUp("You must be in the same voice channel as mine");
     }
 
-    player = message.client.musicManager.create({
+    player = guild.client.musicManager.create({
       guild: guild.id,
       textChannel: channel.id,
       voiceChannel: member.voice.channel.id,
@@ -45,14 +47,14 @@ module.exports = class Play extends Command {
     let res;
 
     try {
-      res = await player.search(query, message.author);
+      res = await player.search(query, user);
       if (res.loadType === "LOAD_FAILED") {
         if (!player.queue.current) player.destroy();
         throw res.exception;
       }
     } catch (err) {
       this.client.logger.error("Search Exception", err);
-      return channel.send("There was an error while searching");
+      return interaction.followUp("There was an error while searching");
     }
 
     let embed = new MessageEmbed().setColor(EMBED_COLORS.BOT_EMBED);
@@ -61,7 +63,7 @@ module.exports = class Play extends Command {
     switch (res.loadType) {
       case "NO_MATCHES":
         if (!player.queue.current) player.destroy();
-        return channel.send(`No results found matching ${query}`);
+        return interaction.followUp(`No results found matching ${query}`);
 
       case "TRACK_LOADED":
         track = res.tracks[0];
@@ -78,8 +80,7 @@ module.exports = class Play extends Command {
           .setFooter(`Requested By: ${track.requester.tag}`);
 
         if (player.queue.totalSize > 0) embed.addField("Position in Queue", (player.queue.size - 0).toString(), true);
-
-        break;
+        return interaction.followUp({ embeds: [embed] });
 
       case "PLAYLIST_LOADED":
         player.queue.add(res.tracks);
@@ -94,7 +95,7 @@ module.exports = class Play extends Command {
           .addField("Playlist duration", "`" + prettyMs(res.playlist.duration, { colonNotation: true }) + "`", true)
           .setFooter(`Requested By: ${res.tracks[0].requester.tag}`);
 
-        break;
+        return interaction.followUp({ embeds: [embed] });
 
       case "SEARCH_RESULT":
         let max = this.client.config.MUSIC.MAX_SEARCH_RESULTS,
@@ -112,30 +113,33 @@ module.exports = class Play extends Command {
           .setDescription(results)
           .setFooter("Please enter song number you wish to add to queue. Type 'end' to cancel");
 
-        channel.send({ embeds: [embed] });
+        await interaction.followUp("Results fetched");
+        await channel.send({ embeds: [embed] });
 
         try {
           collector = await channel.awaitMessages({
-            filter: (m) => m.author.id === message.author.id,
+            filter: (m) => m.author.id === user.id,
             max: 1,
             time: 30e3,
             errors: ["time"],
           });
         } catch (e) {
           if (!player.queue.current) player.destroy();
-          return message.reply("you didn't provide a selection.");
+          return channel.send("you didn't provide a selection.");
         }
 
-        const first = collector.first().content;
+        const sentMsg = collector.first();
+        const content = sentMsg.content;
 
-        if (first.toLowerCase() === "end") {
+        if (content.toLowerCase() === "end") {
           if (!player.queue.current) player.destroy();
-          return channel.send("Cancelled selection.");
+          return sentMsg.reply("Cancelled selection.");
         }
 
-        const index = Number(first) - 1;
-        if (isNaN(index) || index < 0 || index > max - 1)
-          return message.reply(`You must provide a valid number input between (1-${max}).`);
+        const index = Number(content) - 1;
+        if (isNaN(index) || index < 0 || index > max - 1) {
+          return sentMsg.reply(`You must provide a valid number input between (1-${max}).`);
+        }
 
         track = res.tracks[index];
 
@@ -150,8 +154,7 @@ module.exports = class Play extends Command {
           .setFooter(`Requested By: ${track.requester.tag}`);
 
         if (player.queue.totalSize > 0) embed.addField("Position in Queue", (player.queue.size - 0).toString(), true);
+        channel.send({ embeds: [embed] });
     }
-
-    channel.send({ embeds: [embed] });
   }
 };
